requirements.txt
asgiref==3.2.3
Django==3.0
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
